made frame rate a float

+++ v2e/v2e_args.py
@@ -56,9 +56,9 @@ def v2e_args(parser):
     outGroupDvsVideo.add_argument("--dvs_vid_full_scale", type=int, default=2, help="set full scale event count histogram count for DVS videos to be this many ON or OFF events for full white or black.")
     outGroupDvsVideo.add_argument("--output_height",
                         help="height of output DVS data in pixels. If None, same as input video.")
-    outGroupDvsVideo.add_argument("--output_width", 
+    outGroupDvsVideo.add_argument("--output_width",
                         help="width of output DVS data in pixels. If None, same as input video.")
-    outGroupDvsVideo.add_argument("--frame_rate", type=int,
+    outGroupDvsVideo.add_argument("--frame_rate", type=float,
                                   help="implies --dvs_exposure duration 1/framerate.  Equivalent frame rate of --dvs_vid output video; the events will be accummulated as this sample rate; DVS frames will be accumulated for duration 1/frame_rate")
     outGroupDvsVideo.add_argument("--dvs_exposure", nargs='+', type=str, help="mode to finish DVS event integration: duration time: accumulation time in seconds; count n: count n events per frame; area_event N M: frame ends when any area of M x M pixels fills with N events")
